requirements.txt
torch>=1.12
torchvision
tqdm
numpy
Pillow
opencv-python
scikit-image
onnx
onnxruntime
torchmetrics
